Move ESM SSRF test to tests-new and drop fetch route
The fetch route is unrelated to the http.request ESM interception
fix. The test now uses node:test, node:assert, and random ports
matching the tests-new convention.
